+++ R/print.R
@@ -50,7 +50,13 @@ print.clintable <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ...) {
 #' @family Print methods
 #' @rdname print_methods
 #' @export
-knit_print.clintable <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ...) {
+knit_print.clintable <- function(
+  x,
+  n = 3,
+  nrows = 15,
+  apply_defaults = TRUE,
+  ...
+) {
   out <- clintable_as_html(x, n, nrows, apply_defaults, ...)
   knitr::raw_html(out)
 }
@@ -59,7 +65,13 @@ knit_print.clintable <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ..
 #'
 #' Extraction of flextable print method with special handling of clintable pages
 #' @noRd
-clintable_as_html <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ...) {
+clintable_as_html <- function(
+  x,
+  n = 3,
+  nrows = 15,
+  apply_defaults = TRUE,
+  ...
+) {
   refdat <- x$body$dataset
   pg_method <- x$clinify_config$pagination_method
 
@@ -70,7 +82,8 @@ clintable_as_html <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ...) 
   # Apply the default styling
   if (apply_defaults) {
     if (!is.null(titles)) titles <- getOption("clinify_titles_default")(titles)
-    if (!is.null(footnotes)) footnotes <- getOption("clinify_footnotes_default")(footnotes)
+    if (!is.null(footnotes))
+      footnotes <- getOption("clinify_footnotes_default")(footnotes)
     x <- getOption("clinify_table_default")(x)
     if (!is.null(x$clinify_config$footnote_page)) {
       footnote_page <- getOption("clinify_footnotes_default")(footnote_page) |>
@@ -105,36 +118,54 @@ clintable_as_html <- function(x, n = 3, nrows = 15, apply_defaults = TRUE, ...) 
 #' @return Invisible
 #'
 #' @noRd
-print_clinpage <- function(x, titles = NULL, footnotes = NULL, group_label = NULL, captions = NULL) {
+print_clinpage <- function(
+  x,
+  titles = NULL,
+  footnotes = NULL,
+  group_label = NULL,
+  captions = NULL
+) {
   if (!is.null(group_label)) {
     # TODO: Allow formatting on this
-    x <- flextable::add_header_lines(x, values = paste(group_label, collapse = "\n"))
+    x <- flextable::add_header_lines(
+      x,
+      values = paste(group_label, collapse = "\n")
+    )
+    x <- getOption("clinify_grouplabel_default")(x)
     x <- flextable::align(x, 1, 1, "left", part = "header")
   }
   if (!is.null(captions)) {
     # TODO: Allow formatting on this
-    x <- flextable::add_footer_lines(x, values = paste(captions, collapse = "\n"))
+    x <- flextable::add_footer_lines(
+      x,
+      values = paste(captions, collapse = "\n")
+    )
     # This has to be applied after the footer is added
     x <- getOption("clinify_caption_default")(x)
   }
 
   body <- flextable::htmltools_value(x = x)
   # Two different type of leading spaces that appear in the HTML
-  # body[[3]] <- gsub("(?<!th|<td)  ", "&nbsp; ", body[[3]], perl = TRUE)
-  body[[3]] <- gsub("(?<!th)  ", "&nbsp; ", body[[3]], perl = TRUE)
+  body[[3]] <- gsub("(?<!th|<td)  ", "&nbsp; ", body[[3]], perl = TRUE)
+  # body[[3]] <- gsub("(?<!th)  ", "&nbsp; ", body[[3]], perl = TRUE)
   body[[3]] <- gsub("(<span\\b[^>]*>) ", "\\1&nbsp;", body[[3]], perl = TRUE)
   # Concurrent spaces
   body[[3]] <- gsub("&nbsp;  ", "&nbsp;&nbsp; ", body[[3]], perl = TRUE)
 
   if (!is.null(titles)) {
-    # TODO: This should take into consideration how many cells are merged within the header
-    titles <- flextable::width(titles, width = flextable::flextable_dim(x)$widths / 2)
+    titles <- flextable::width(
+      titles,
+      width = flextable::flextable_dim(x)$widths / 2
+    )
     hdr <- flextable::htmltools_value(x = titles)[[3]]
     body[[3]] <- htmltools::HTML(paste0(hdr, body[[3]]))
   }
 
   if (!is.null(footnotes)) {
-    footnotes <- flextable::width(footnotes, width = flextable::flextable_dim(x)$widths / 2)
+    footnotes <- flextable::width(
+      footnotes,
+      width = flextable::flextable_dim(x)$widths / 2
+    )
     ftr <- flextable::htmltools_value(x = footnotes)[[3]]
     body[[3]] <- htmltools::HTML(paste0(body[[3]], ftr))
   }
@@ -147,7 +178,13 @@ print_clinpage <- function(x, titles = NULL, footnotes = NULL, group_label = NUL
 #' @param x a clintable object
 #' @param n number of pages within the clintable to print
 #' @noRd
-print_alternating <- function(x, n, titles = NULL, footnotes = NULL, footnote_page = NULL) {
+print_alternating <- function(
+  x,
+  n,
+  titles = NULL,
+  footnotes = NULL,
+  footnote_page = NULL
+) {
   pag_idx <- x$clinify_config$pagination_idx
 
   # Don't try to print more pages than requested
@@ -201,12 +238,20 @@ print_alternating <- function(x, n, titles = NULL, footnotes = NULL, footnote_pa
   for (i in seq_along(pgs)) {
     # Generate the divs of the individual pages
     if (i == 1) {
-      args <- append(args, list(div(id = sprintf("page%s", i), class = "page", pgs[i])))
+      args <- append(
+        args,
+        list(div(id = sprintf("page%s", i), class = "page", pgs[i]))
+      )
     } else {
       # Only display first page
       args <- append(
         args,
-        list(div(id = sprintf("page%s", i), style = "display:none;", class = "page", pgs[i]))
+        list(div(
+          id = sprintf("page%s", i),
+          style = "display:none;",
+          class = "page",
+          pgs[i]
+        ))
       )
     }
 
@@ -231,10 +276,12 @@ print_alternating <- function(x, n, titles = NULL, footnotes = NULL, footnote_pa
 
 #' JS for the page buttons
 #' @noRd
-pagination_js <- tags$script(HTML("
+pagination_js <- tags$script(HTML(
+  "
   function showPage(pageNum) {
     var pages = document.querySelectorAll('.page');
     pages.forEach(page => page.style.display = 'none');
     document.getElementById('page' + pageNum).style.display = 'block';
   }
-"))
+"
+))